+++ ptai-teamcity-plugin/ptai-teamcity-plugin-agent/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/teamcity/agent/operations/TeamcityAstOperations.java
@@ -40,7 +40,7 @@ public class TeamcityAstOperations implements AstOperations {
     }
 
     @Override
-    public void scanCompleteCallback(@NonNull ScanBrief scanBrief) throws GenericException {
+    public void scanCompleteCallback(ScanBrief scanBrief) throws GenericException {
 
     }